requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
Flask-Migrate==4.0.5
requests==2.31.0
Flask-APScheduler==1.12.3
